requirements.txt
streamlit
requests
python-dotenv
stqdm
openpyxl